forgot about review CRUD.

+++ api/src/routes/reviews/index.ts
@@ -1,7 +1,7 @@
+import { Sequelize } from 'sequelize';
 import { dbConnect, models } from '../../db';
 
 interface ReviewData {
-  cocktailId: number;
   rating: number;
   body?: string;
   scaleComposition?: number;
@@ -24,10 +24,14 @@ const getReviewsForCocktail = async (cocktailId: string) => {
   });
 };
 
-const addReview = async (reviewData: ReviewData, userId: number) => {
+const testingFunction = async (cocktailId: string) => {
+  return await refreshCocktailStats(123, +cocktailId);
+}
+
+const addReview = async (cocktailId: string, reviewData: ReviewData, userId: number) => {
   await dbConnect();
 
-  const { cocktailId, rating, body, scaleComposition, scaleSpirited } = reviewData;
+  const { rating, body, scaleComposition, scaleSpirited } = reviewData;
 
   // confirm the user doesn't have a list already with the same name
   const existingReview = await models.rating.findOne({
@@ -41,7 +45,7 @@ const addReview = async (reviewData: ReviewData, userId: number) => {
     throw new Error('review from user for that cocktail already exists!');
   }
 
-  await models.rating.create({
+  const review = await models.rating.create({
     cocktail_id: cocktailId,
     user_id: userId,
     rating,
@@ -49,6 +53,14 @@ const addReview = async (reviewData: ReviewData, userId: number) => {
     scale_composition: scaleComposition,
     scale_spirited: scaleSpirited,
   });
+
+  // @ts-ignore
+  const cocktailUpdateResult = await refreshCocktailStats(review.id, +cocktailId);
+
+  return {
+    ...review.dataValues,
+    cocktailUpdateStatus: cocktailUpdateResult,
+  };
 }
 
 const updateReview = async (reviewId: string, reviewData: ReviewData, userId: number) => {
@@ -56,7 +68,7 @@ const updateReview = async (reviewId: string, reviewData: ReviewData, userId: nu
 
   const { rating, body, scaleComposition, scaleSpirited } = reviewData;
 
-  await models.rating.update({
+  const updateResult = await models.rating.update({
     rating,
     body,
     scale_composition: scaleComposition,
@@ -68,17 +80,91 @@ const updateReview = async (reviewId: string, reviewData: ReviewData, userId: nu
       user_id: userId,
     }
   });
+
+  const cocktailUpdateResult = await refreshCocktailStats(+reviewId);
+
+  return {
+    reviewsUpdated: updateResult,
+    cocktailUpdateStatus: cocktailUpdateResult,
+  };
 }
 
 const deleteReview = async (reviewId: string, userId: number) => {
   await dbConnect();
 
-  await models.rating.destroy({
+  // need to grab the cocktail id before we delete the review
+  const existingReview = await models.rating.findByPk(reviewId);
+
+  if (!existingReview) {
+    throw new Error('no existing review!');
+  }
+
+  // @ts-ignore
+  const cocktailId = existingReview.cocktail_id;
+
+  const destroyResult = await models.rating.destroy({
     where: {
       id: reviewId,
       user_id: userId,
     }
   });
+
+  const cocktailUpdateResult = await refreshCocktailStats(+reviewId, cocktailId);
+
+  return {
+    reviewsDeleted: destroyResult,
+    cocktailUpdateStatus: cocktailUpdateResult,
+  };
+}
+
+const refreshCocktailStats = async (reviewId: number, cocktailId?: number) => {
+  let finalCocktailId = cocktailId;
+  if (!finalCocktailId) {
+    const existingReview = await models.rating.findByPk(reviewId);
+
+    if (!existingReview) {
+      throw new Error('missing cocktail id and existing review to look up cocktail id!');
+    }
+
+    // @ts-ignore
+    finalCocktailId = existingReview.cocktail_id;
+  }
+
+  const statsForCocktails = await models.rating.findAll({
+    where: {
+      cocktail_id: finalCocktailId,
+    },
+    attributes: [
+      'cocktail_id',
+      [Sequelize.fn('avg', Sequelize.col('rating')),'avgRating'],
+      [Sequelize.fn('avg', Sequelize.col('scale_composition')),'avgComposition'],
+      [Sequelize.fn('avg', Sequelize.col('scale_spirited')),'avgSpirited'],
+    ],
+    group: ['cocktail_id'],
+  });
+
+  console.log(statsForCocktails);
+
+  const statsForCocktail = statsForCocktails[0].dataValues;
+
+  const cocktailUpdate = await models.cocktail.update({
+    // @ts-ignore
+    avg_rating: statsForCocktail.avgRating,
+    // @ts-ignore
+    avg_spirited: statsForCocktail.avgSpirited,
+    // @ts-ignore
+    avg_composition: statsForCocktail.avgComposition,
+    updated_at: Date.now(),
+  }, {
+    where: {
+      id: finalCocktailId,
+    }
+  });
+
+  return {
+    rowsUpdated: cocktailUpdate,
+    ...statsForCocktail,
+  };
 }
 
 export default {
@@ -86,4 +172,5 @@ export default {
   addReview,
   updateReview,
   deleteReview,
+  testingFunction,
 }
